+++ CHANGES.rst
@@ -5,7 +5,7 @@
 -----------------
 
 - Fixed attributeError when loading a QGIS project (#1063)
-
+- Fix in Rasters to NetCDF algorithm to properly convert the units Enum to string (#1067)
 
 3.11 (2024-11-12)
 -----------------