+++ CHANGES.rst
@@ -5,7 +5,8 @@
 -----------------
 
 - Added Processing Algorithm "Extract structure control actions" (#926)
-
+- Fixed attributeError when loading a QGIS project (#1063)
+- Fix in Rasters to NetCDF algorithm to properly convert the units Enum to string (#1067)
 
 3.11 (2024-11-12)
 -----------------