+++ CHANGES.rst
@@ -6,6 +6,7 @@
 
 - Compatibility with Python 3.12 (#1061)
 - Bumped h5py to 3.10.0 and scipy to 1.13.0 for python 3.12 compatibility (#1061)
+- Added Processing Algorithm "Extract structure control actions" (#926)
 - Fixed attributeError when loading a QGIS project (#1063)
 - Fix in Rasters to NetCDF algorithm to properly convert the units Enum to string (#1067)