+++ CHANGES.rst
@@ -1,8 +1,8 @@
 3Di Results Analysis changelog
 ========================
 
-3.12 (unreleased)
------------------
+3.11.1 (2024-12-06)
+-------------------
 
 - Compatibility with Python 3.12 (#1061)
 - Bumped h5py to 3.10.0 and scipy to 1.13.0 for python 3.12 compatibility (#1061)